extmods: add utils directories in sys.path
When the minion start clean and there are not utils directory in
the extmods cache, any import from a module to any code inside the
utils directory will fail. A second run of the highstate will fix
this issue, as the utils directories are added into sys.path inside
config/__init__.py, as part of the master and minion startup.

This commit add the utils directories during the synchronization of
the custom modules.

Fixes #51958